+++ sqflite/lib/src/exception_impl.dart
@@ -2,6 +2,7 @@ import 'package:flutter/services.dart';
 import 'package:sqflite/src/constant.dart';
 import 'package:sqflite/src/exception.dart';
 
+/// Wrap any exception to a [DatabastException]
 Future<T> wrapDatabaseException<T>(Future<T> action()) async {
   try {
     final T result = await action();