+++ sqflite/lib/utils/utils.dart
@@ -1,6 +1,6 @@
-import 'package:sqflite/src/constant.dart' as constant;
 import 'package:sqflite/src/sqflite_impl.dart';
 import 'package:sqflite/src/utils.dart';
+import 'package:sqflite/src/utils.dart' as impl;
 
 /// helper to get the first int value in a query
 /// Useful for COUNT(*) queries
@@ -27,13 +27,32 @@ String hex(List<int> bytes) {
   return buffer.toString().toUpperCase();
 }
 
-Duration lockWarningDuration = constant.lockWarningDuration;
-void Function() lockWarningCallback = () {
-  print('Warning database has been locked for $lockWarningDuration. '
-      'Make sure you always use the transaction object for database operations during a transaction');
-};
+/// Deprecated since 1.1.7+.
+///
+/// Used internally.
+@deprecated
+void Function() get lockWarningCallback => impl.lockWarningCallback;
 
+/// Deprecated since 1.1.7+.
+@deprecated
+set lockWarningCallback(void Function() callback) =>
+    impl.lockWarningCallback = callback;
+
+/// Deprecated since 1.1.7+.
+@deprecated
+Duration get lockWarningDuration => impl.lockWarningDuration;
+
+/// Deprecated since 1.1.7+.
+@deprecated
+set lockWarningDuration(Duration duration) =>
+    impl.lockWarningDuration = duration;
+
+/// Change database lock behavior mechanism.
+///
+/// Default behavior is to print a message if a command hangs for more than
+/// 10 seconds. Set en empty callback (not null) to prevent it from being
+/// displayed.
 void setLockWarningInfo({Duration duration, void callback()}) {
-  lockWarningDuration = duration ?? lockWarningDuration;
-  lockWarningCallback = callback ?? lockWarningCallback;
+  impl.lockWarningDuration = duration ?? impl.lockWarningDuration;
+  impl.lockWarningCallback = callback ?? impl.lockWarningCallback;
 }